TheSkyX tracking and Float64/32 platesolve issue

+++ devices/camera.py
@@ -2625,6 +2625,7 @@ class Camera:
                                 # Interpolate to make a high resolution version for focussing
                                 # and platesolving
                                 hdufocus.data=demosaicing_CFA_Bayer_bilinear(hdufocus.data, 'RGGB')[:,:,1]
+                                hdufocus.data=hdufocus.data.astype("float32")
                                 
                                 
 
@@ -2635,7 +2636,7 @@ class Camera:
                         focusimg = np.asarray(
                             hdufocus.data
                         )  # + 100   #maintain a + pedestal for sep  THIS SHOULD not be needed for a raw input file.
-                        focusimg = focusimg.astype("float")
+                        focusimg = focusimg.astype("float32")
                         focusimg = focusimg.copy(
                             order="C"
                         )  # NB Should we move this up to where we read the array?
